# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OSShip open-source mentorship platform API exposed via the gateway.",
        "title": "OSShip API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/listings": {
            "get": {
                "description": "Returns mentorship listings filtered by status and optional OSS project.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List listings",
                "parameters": [
                    {
                        "type": "string",
                        "default": "active",
                        "description": "Listing status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by OSS project name",
                        "name": "oss_project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apidoc.Listing"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a mentorship listing (approved mentor only).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Create listing",
                "parameters": [
                    {
                        "description": "Listing payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apidoc.CreateListingReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apidoc.Listing"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apidoc.CreateListingReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration_weeks": {
                    "type": "integer",
                    "example": 8
                },
                "oss_project_name": {
                    "type": "string",
                    "example": "kubernetes"
                },
                "oss_repo_url": {
                    "type": "string",
                    "example": "https://github.com/kubernetes/kubernetes"
                },
                "price_cents": {
                    "type": "integer",
                    "example": 9900
                },
                "total_slots": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "apidoc.ErrorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
...
        "apidoc.Listing": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration_weeks": {
                    "type": "integer",
                    "example": 8
                },
                "filled_slots": {
                    "type": "integer",
                    "example": 2
                },
                "id": {
                    "type": "string"
                },
                "mentor_display_name": {
                    "type": "string"
                },
                "mentor_github_username": {
                    "type": "string"
                },
                "mentor_id": {
                    "type": "string"
                },
                "oss_project_name": {
                    "type": "string",
                    "example": "kubernetes"
                },
                "oss_repo_url": {
                    "type": "string",
                    "example": "https://github.com/kubernetes/kubernetes"
                },
                "price_cents": {
                    "type": "integer",
                    "example": 9900
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "total_slots": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. Example: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}